requirements.txt
# LLM Semantic Book Recommender Dependencies
flask==3.1.0
pandas==2.2.2
numpy==1.26.4
langchain==0.1.12
langchain_community==0.0.29
langchain_openai==0.0.8
langchain_text_splitters==0.0.1
langchain_chroma==0.1.1
chromadb==0.5.23
openai==1.30.0
python-dotenv==1.0.1
Werkzeug==3.1.3
blinker==1.9.0
itsdangerous==2.2.0